Port to MSVC/Windows
The encode and decode packet tests do not work currently as the compiler does not like the test harnish.

connect - publish and connect - subscribe work though and seem to not crash for now.
It is possible that the mqtt.dll is not working correctly right now, possibly the exports are not right.

Re #21